+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Enforce that the entire marker string is parsed (:issue:`687`)
 * Requirement parsing no longer automatically validates the URL (:issue:`120`)
+* Canonicalize names for requirements comparison (:issue:`644`)
 
 23.1 - 2023-04-12
 ~~~~~~~~~~~~~~~~~